The main purpose of this program is to abstractly simulate the reassembly of IP fragments.
This program accepts a input file with IP fragment information as an
input parameter. The expected behaviour of the program is the following:

For every "IP source address, ID" combination, the program reads and
prints whether the fragments given form a complete datagram. If that is
true, it  prints the Source IP address in dotted decimal notation, ID
and the total length of the datagram. If not, it prints the Source IP
address and ID and the error encountered while reassembling such as,
"Duplicate Fragment", "Missing Fragment", "Mismatch in length" etc..